init 转换为百度墨卡托坐标

+++ food-interface/src/main/java/com/accendl/food/dto/MapLocationDTO.java
@@ -4,14 +4,13 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 @Data
 @Builder
 public class MapLocationDTO implements Serializable {
     private Integer id;
 
-    private BigDecimal mercatorX;
+    private double mercatorX;
 
-    private BigDecimal mercatorY;
+    private double mercatorY;
 }
